Added finishing touches
